// src/enums/general-message-code.enum.ts
export enum GeneralMessageCodeEnum {
  NotFound = "NotFound",
  InternalServerError = "InternalServerError",
  BadData = "BadData",
  MethodNotAllowed = "MethodNotAllowed",
  Forbidden = "Forbidden",

  UserRegistrationSuccess = "UserRegistrationSuccess",
  UserVerificationEmailSentSuccess = "UserVerificationEmailSentSuccess",
  UserEmailVerificationSuccess = "UserEmailVerificationSuccess",

  AssetSuccessfullyUploaded = "AssetSuccessfullyUploaded",
  LoginSuccess = "LoginSuccess",
  AdminUserCreateSuccess = "AdminUserCreateSuccess",
  AdminUserListSuccess = "AdminUserListSuccess",
  AdminUserLoginSuccess = "AdminUserLoginSuccess",
}